legacy_container_UX/backend/requirements.txt
# Core Framework
fastapi>=0.120.1
starlette>=0.49.1
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic[email]
uvicorn[standard]==0.37.0
python-multipart

# Environment
python-dotenv==1.1.1

# Observability
logfire[fastapi]

# GCP & Database
google-cloud-firestore==2.21.0
google-cloud-run==0.10.11
google-cloud-storage>=2.14.0

# Redis Cache
redis[hiredis]==5.0.1

# Authentication
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
passlib==1.7.4
google-auth-oauthlib>=1.2.0

# HTTP Client
httpx==0.28.1

# MCP Server
mcp>=1.10.0

# AI Agent Framework
pydantic-ai

# Development
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==5.0.0
pytest-xdist==3.6.1
schemathesis==3.36.2
ruff==0.8.3
mypy==1.11.0

# Pre-commit & Security
pre-commit==4.0.1
bandit[toml]==1.7.10
pip-audit==2.7.3
